+++ bill/invoice_test.go
@@ -1050,9 +1050,9 @@ func TestInvoiceForUnknownRegime(t *testing.T) {
 func TestNormalization(t *testing.T) {
 	inv := baseInvoiceWithLines(t)
 	inv.Series = " bar 2024 "
-	inv.Code = " 123 Test "
+	inv.Code = " 123_Test "
 	require.NoError(t, inv.Calculate())
-	assert.Equal(t, cbc.Code("BAR-2024"), inv.Series)
+	assert.Equal(t, cbc.Code("BAR 2024"), inv.Series)
 	assert.Equal(t, cbc.Code("123-TEST"), inv.Code)
 }